+++ src/components/signup/CheckInputField.jsx
@@ -1,38 +1,48 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { InforEach } from './SingupForm';
+import { InforEach, Label, InputField } from './SingupForm';
 import { useCheckValidity } from '../../api/signupApi';
 
 const EMAIL_REG = /^\w+@[a-zA-Z_]+?\.[a-zA-Z]{2,3}$/;
 
 export default function CheckInputField({ title, name, setUserInfor }) {
   const [value, setValue] = useState('');
-  const [validty, setValidty] = useState(false);
+  const [validity, setValidity] = useState(false);
   const { refetch } = useCheckValidity(name, value);
 
   const checkEmailValidity = () => {
-    if (EMAIL_REG.text(value)) {
+    if (!EMAIL_REG.test(value)) {
       alert('유효하지 않는 형식의 이메일입니다');
-      return;
+      setValidity(false);
+      return false;
+    } else {
+      return true;
     }
   };
 
   const checkIdValidity = () => {
     if (value.length < 8) {
       alert('아이디는 최소 8글자여야 합니다.');
-      return;
+      setValidity(false);
+      return false;
+    } else {
+      return true;
     }
   };
 
   const setInfor = async () => {
     const data = await refetch();
-    console.log(data.data.data.data.exists);
+
     if (data.data.data.data.exists) {
       alert(`이미 존재하는 ${title}입니다.`);
+      setValidity(false);
+      setUserInfor(`is${name}Unique`, false);
     } else {
       alert(`사용 가능한 ${title}입니다`);
-      setUserInfor(name, value);
+      setUserInfor(`is${name}Unique`, true);
+      setValidity(true);
     }
+    setUserInfor(name, value);
   };
 
   return (
@@ -46,13 +56,15 @@ export default function CheckInputField({ title, name, setUserInfor }) {
       <RepetitionCheckBtn
         onClick={async (e) => {
           e.preventDefault();
+          let result = false;
           if (name === 'email') {
-            checkEmailValidity();
+            result = checkEmailValidity();
           } else if (name === 'loginId') {
-            checkIdValidity();
+            result = checkIdValidity();
           }
-
-          await setInfor();
+          if (result) {
+            await setInfor();
+          }
         }}
       >
         중복 확인
@@ -60,30 +72,6 @@ export default function CheckInputField({ title, name, setUserInfor }) {
     </InforEach>
   );
 }
-
-const Label = styled.div`
-  font-size: 14px;
-  color: #606060;
-  padding-bottom: 15px;
-`;
-
-const InputField = styled.input.attrs({ requried: true })`
-  border: none;
-  border-bottom: 2px solid #f0f0f0;
-  width: '238px';
-  height: 16px;
-  padding-bottom: 8px;
-
-  &:focus {
-    outline: none;
-  }
-
-  &::-ms-reveal,
-  ::-ms-clear {
-    display: none;
-  }
-`;
-
 const RepetitionCheckBtn = styled.button`
   position: absolute;
   transform: translate(260px, 10px);
